added A2 to A
