--- pages/A2.js
@@ -1,156 +1,0 @@
-// @flow
-
-import React, { Component } from "react";
-
-import { Grid } from "material-ui";
-
-import withRedux from "next-redux-wrapper";
-import { initStore, loadDataStore } from "../store";
-
-import { withI18next } from "../lib/withI18next";
-import Layout from "../components/layout";
-import { logEvent } from "../utils/analytics";
-import Link from "next/link";
-import SelectButton from "../components/select_button";
-import { bindActionCreators } from "redux";
-import { fetchFromAirtable } from "../utils/airtable";
-
-type Props = {
-  i18n: mixed,
-  t: mixed,
-  storeHydrated: boolean,
-  loadDataStore: mixed,
-  patronTypes: mixed,
-  url: mixed
-};
-
-export class App extends Component<Props> {
-  props: Props;
-
-  constructor() {
-    super();
-    this.state = {
-      patronTypes: [],
-      selectedOptions: []
-    };
-  }
-
-  toggleButton = id => {
-    let selected = this.state.selectedOptions;
-    const index = selected.indexOf(id);
-    if (index >= 0) {
-      selected.splice(index, 1);
-    } else {
-      selected.push(id);
-    }
-    this.setState({
-      selectedOptions: selected
-    });
-  };
-
-  changeLanguage = () => {
-    this.props.i18n.changeLanguage(this.props.t("other-language-code"));
-    logEvent("Language change", this.props.t("other-language"));
-  };
-
-  async componentWillMount() {
-    if (!this.props.storeHydrated) {
-      fetchFromAirtable(this.props.loadDataStore);
-    }
-  }
-
-  render() {
-    const { i18n, t } = this.props; // eslint-disable-line no-unused-vars
-
-    return (
-      <Layout i18n={i18n} t={t}>
-        <div style={{ padding: 12 }}>
-          <Grid container spacing={24}>
-            <Grid item xs={12}>
-              <p style={{ textAlign: "center", fontSize: "1.5em" }}>
-                {t("A2.What best describes your status?")}
-              </p>
-            </Grid>
-          </Grid>
-
-          {this.props.patronTypes.map((type, i) => (
-            <Grid
-              container
-              key={i}
-              justify="center"
-              spacing={24}
-              style={{ marginTop: "1em" }}
-            >
-              <Grid item sm={4} xs={12}>
-                <SelectButton
-                  id={type.id}
-                  text={
-                    t("current-language-code") === "en"
-                      ? type.name_en
-                      : type.name_fr
-                  }
-                  onClick={this.toggleButton}
-                  isDown={this.state.selectedOptions.indexOf(type.id) >= 0}
-                />
-              </Grid>
-            </Grid>
-          ))}
-
-          <Grid
-            container
-            justify="center"
-            spacing={24}
-            style={{ marginTop: "3em" }}
-          >
-            <Grid item sm={4} xs={12}>
-              <SelectButton
-                text={t("A2.See Results")}
-                href={
-                  "A3?lng=" +
-                  t("current-language-code") +
-                  "&benefitTypes=" +
-                  this.props.url.query.benefitTypes +
-                  "&patronTypes=" +
-                  this.state.selectedOptions.join()
-                }
-                isDown={false}
-              />
-            </Grid>
-          </Grid>
-
-          <Grid
-            container
-            justify="center"
-            spacing={24}
-            style={{ marginTop: "1em" }}
-          >
-            <Grid item sm={4} xs={12}>
-              <p style={{ textAlign: "center", fontSize: "1em" }}>
-                <Link href="all-benefits">
-                  <a>{t("Show All Benefits")}</a>
-                </Link>
-              </p>
-            </Grid>
-          </Grid>
-        </div>
-      </Layout>
-    );
-  }
-}
-
-const mapDispatchToProps = dispatch => {
-  return {
-    loadDataStore: bindActionCreators(loadDataStore, dispatch)
-  };
-};
-
-const mapStateToProps = state => {
-  return {
-    storeHydrated: state.storeHydrated,
-    patronTypes: state.patronTypes
-  };
-};
-
-export default withRedux(initStore, mapStateToProps, mapDispatchToProps)(
-  withI18next()(App)
-);
